# docs/swagger.yaml
basePath: /api/v1/
definitions:
  v1.CategoryResponse:
    properties:
      id:
        type: integer
      label:
        type: string
    type: object
...
  v1.CreateCategoryRequest:
    properties:
      label:
        type: string
    type: object
...
  v1.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  v1.LoginResponse:
    properties:
      token:
        description: bearer token
        type: string
    type: object
...
info:
  contact:
    email: kerhael.me@gmail.com
    name: API Support
  description: API de suivi de comptabilité
  termsOfService: https://example.com/terms/
  title: Accounting API
  version: "1.0"
paths:
  /categories/:
    get:
      consumes:
      - application/json
      description: Retrieve all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.CategoryResponse'
            type: array
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/v1.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.CategoryResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Retrieve a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.CategoryResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a category
      tags:
      - categories
...
  /users/login:
    post:
      consumes:
      - application/json
      description: User login. A rate limiter prevents from brute force attacks (speed
        1s, burst 5)
      parameters:
      - description: Login payload
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/v1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.LoginResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "429":
          description: Too many requests error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Login
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"